+++ tests/test_structure.py
@@ -5,6 +5,7 @@ import pytest
 import playa
 from playa.exceptions import PDFEncryptionError
 from playa.page import Annotation, XObjectObject, TextObject
+from playa.pdftypes import BBOX_NONE
 from playa.structure import Element, Tree, ContentItem, ContentObject
 
 from .data import ALLPDFS, CONTRIB, PASSWORDS, TESTDIR, XFAILS
@@ -71,6 +72,7 @@ def test_annotations() -> None:
             for kid in link:
                 if isinstance(kid, ContentObject):
                     assert isinstance(kid.obj, Annotation)
+                    assert kid.bbox is not BBOX_NONE
 
 
 def test_content_xobjects() -> None:
@@ -98,15 +100,30 @@ def test_structure_bbox() -> None:
         assert pdf.structure is not None
         table = pdf.structure.find("Table")
         assert table is not None
-        print(table.bbox)
+        assert table.bbox is not BBOX_NONE
         li = pdf.structure.find("LI")
         assert li is not None
-        print(li.bbox)
+        assert li.bbox is not BBOX_NONE
+        for item in li.contents:
+            assert item.bbox is not BBOX_NONE
     with playa.open(TESTDIR / "image_structure.pdf") as pdf:
         assert pdf.structure is not None
         figure = pdf.structure.find("Figure")
         assert figure is not None
-        print(figure.bbox)
+        assert figure.bbox is not BBOX_NONE
+        for item in figure.contents:
+            assert item.bbox is not BBOX_NONE
+
+
+def test_content_structure() -> None:
+    """Verify that we can access structure elements from content objects."""
+    with playa.open(TESTDIR / "pdf_structure.pdf") as pdf:
+        for obj in pdf.pages[0]:
+            if obj.object_type == "path":
+                assert obj.parent is None
+            else:
+                assert obj.parent is not None
+                assert obj.parent.role in ("P", "H1", "H2", "H3")
 
 
 if __name__ == "__main__":